Rename Common.h -> QueryId.h
