+++ src/engine/QueryExecutionContext.h
@@ -22,7 +22,7 @@
 #include "util/ConcurrentCache.h"
 #include "util/Log.h"
 #include "util/Synchronized.h"
-#include "util/http/websocket/Common.h"
+#include "util/http/websocket/QueryId.h"
 
 using std::shared_ptr;
 using std::string;
