+++ src/util/http/websocket/WebSocketSession.cpp
@@ -10,8 +10,8 @@
 #include <optional>
 
 #include "util/http/HttpUtils.h"
-#include "util/http/websocket/Common.h"
 #include "util/http/websocket/QueryHub.h"
+#include "util/http/websocket/QueryId.h"
 #include "util/http/websocket/UpdateFetcher.h"
 
 namespace ad_utility::websocket {
